Fix Bloom build pipeline
We can't have the build job trigger (and wait for) the unit test
job because this keeps the build job running, blocking the unit test
job from running on the same node (which we told it to do). This
change changes the behavior to be the same as for pre-merge builds.

Tried to use the Join plugin, but that the BuildPipeline plugin
doesn't properly work with that and shows a build for each of the
build waiting to be joined, but only reports one of these builds to
as being run, the others as pending. This looks ugly, so I abandoned
that approach.

Change-Id: Ie25006661f38a732f0b08673720ca89cb1c588bf