Fix problem with proxy
We should check that http_proxy is not set before setting it to
a potentially unset HTTP_PROXY variable.

In my testing before HTTP_PROXY was always set and http_proxy wasn't
but now it seems it is the other way round - maybe because of an
updated Jenkins plugin. In either case it's a good idea to check if
the variable is null before trying to overwrite it.

Change-Id: Iad0d436e972a26a9283e40f88fcfffe796d357d6