Allow to set location of pbuilder tools
This change allows to set PBUILDER_TOOLS_PATH environment variable
to specify the path to the pbuilder tools. If not set the tools are
assumed to be in $HOME/FwSupportTools/packaging/pbuilder.

Also allow to set the variable PBUILDERDIR to point to the pbuilder
directory (default: $HOME/pbuilder). While the default values work
for the packaging build agent they may not match the environment
on a developer's machine.

Change-Id: I22872d56b71540ada9f768941de9576da6d15931