[lfmerge] Fix lfmerge-fdo package building job
Currently the multiple-scm plugin doesn't work together with
the gerrit trigger. The gerrit trigger sets environment variable
which the build checks, but if we have multiple scms the other
scms won't find the SHA1 and so the build fails. It only works
when manually triggered.

This change removes the gerritTrigger strategy so that we always
build HEAD, which should be fine for this job.

Change-Id: I7210d6ed09c1720dcc83c4e6a6e683b262a58529