New build job for packaging LfMerge live branch
This one does NOT automatically push the built .deb file to the server,
nor does up automatically upload the .deb file to LLSO. It is intended
to simply publish the .deb file as a build artifact. Then a developer
will download that .deb file and install it on the live server. (This
might happen semi-automatically with a TeamCity build, but that build
will be manually triggered.) So the build configuration on Jenkins
should be set to "Release" rather than "Nightly" for this job.

Change-Id: I11e5524b232b03d1c8306af97d18b21db0602a65