Add QA branch to LfMerge build; upload to proposed
The -proposed suite in Debian is used as a staging ground for packages
that are thought to be ready to release to the stable distribution, but
need a little bit more testing. In other words, release candidates.

The way this patch is written, builds of the "qa" branch would be
automatically uploaded to llso-proposed. If that is not desired, change
the "--suite-name=proposed" line to "--no-upload" in LfMergeJobs.groovy,
and then a developer would need to download the packages by hand, sign
them, and upload them to the -proposed suite.

Change-Id: Iae2dd480b8f3e101aeef9fead468ea72adaff3a4