[lfmerge] Add QA branch, uploading to -proposed
The -proposed suite in Debian is used as a staging ground for packages
that are thought to be ready to release to the stable distribution, but
need a little bit more testing. In other words, release candidates.

The way this patch is written, builds of the "qa" branch would be
automatically uploaded to llso-proposed. If that is not desired, change
the "--suite-name=proposed" line to "--no-upload" in LfMergeJobs.groovy,
and then a developer would need to download the packages by hand, sign
them, and upload them to the -proposed suite.

Change-Id: I7f77f60b255fdb2bd4d6d84c60f009d9f26aea0c