[Packaging] Use `unbuffer` when running sbuild
Usually when building on Jenkins, sbuild detects that there is no
tty connected, so it writes its output only to the log file. This
makes it impossible to see what's going on by looking at the
Jenkins build log. This change uses `unbuffer` to run sbuild which
causes the output to appear in the build log.

Change-Id: I0f75c3519f42772fefac0461fa0949cdc7c190cb