[Keyman] Do trigger pipeline job on scan
The job itself then determines if it should run a complete build.
Because the generic trigger code is in the pipeline job it is
necessary for a new branch to build once so that the generic trigger
knows what to listen for. If the pipeline job gets triggered by
a scan it will exit after executing the generic trigger configuration
code, otherwise it will do a complete build.

Change-Id: I9b952db9faf4be1e18d0f783559288e75e1c494e