[Bloom] Disable master branch build and limit to Xenial
Bloom packaging currently seems to have problems when running on
Bionic, so we limit the packaging agents to the Xenial ones.
Also, the option to trigger a default branch build after a non-default
build is confusing and it's not clear when we would need that, so
this change disables this option.

Change-Id: I34eea454656627d6e12ade6f2838168fa2a75ca3