[scripts] Don't delete line 10
It's unclear why we had added deleting line 10 in the past. Nowadays
however it causes problems, e.g. it deletes the Distribution field
or any other random field that happens to be at line 10 so that
processing the package on llso fails.

This change removes deleting line 10 - we'll see what other problems
this will cause.

Change-Id: I719fd535864a1e86409078baca7d91c510343823