Rename PackageVersion env var: conflcit with NuGet
Recent versions of NuGet apparently read and parse an envrionment
variable named PackageVersion, expecting it to be in NuGet package
version format. When the Debian package version string ends up in that
environment variable with a tilde (~) in it, NuGet complains and the
build errors out. By renaming the PackageVersion variable in the LfMerge
build to DebPackageVersion, we can avoid this env var name conflict.

Change-Id: I8d7790cbc5bef9ccdd394f3685ed533b7ca8836c